+++ src/Frontend/Dfe.Complete/Constants/RouteConstants.cs
@@ -50,6 +50,9 @@
         public const string ProjectAbout = Project + "/information";
         public const string ProjectEditAbout = ProjectAbout + "/edit";
         public const string ProjectDateHistory = Project + "/date-history";
+        public const string ChangeProjectDateHistory = Project + "/date-history/new";
+        public const string ChangeProjectDateHistoryReason = Project + "/date-history/reason";
+        public const string ChangeProjectDateHistoryConfirm = Project + "/date-history/reasons/later";
         
         public const string ProjectHandoverWithDeliveryOfficerTask = ProjectTaskList + "/handover";
         public const string ProjectStakeholderKickoffTask = ProjectTaskList + "/stakeholder-kickoff";