+++ src/Frontend/Dfe.Complete/Pages/Projects/DateHistory/ConfirmDateChange.cshtml.cs
@@ -1,0 +1,16 @@
+using Dfe.Complete.Pages.Projects.ProjectView;
+using MediatR;
+using Microsoft.AspNetCore.Mvc;
+
+namespace Dfe.Complete.Pages.Projects.DateHistory
+{
+    
+    public class ConfirmDateChangeModel(ISender sender) : ProjectLayoutModel(sender, ConversionDateHistoryNavigation)
+    {
+        public override async Task<IActionResult> OnGet()
+        {
+            await base.OnGet();
+            return Page();
+        }
+    }
+}